Reorganize packages
- Fold workspace/provision/ideUrl.go into workspace/status.go
- Move rest-apis related files into workspace/restapis package
- Set che-rest-apis image in configmap
