+++ deploy/controller_config.yaml
@@ -8,6 +8,6 @@ data:
   ingress.global.domain: 192.168.99.100.nip.io
   plugin.registry.url: http://che-plugin-registry.192.168.99.100.nip.io/v3
   che.workspace.plugin_broker.artifacts.image: quay.io/eclipse/che-plugin-artifacts-broker:v3.1.0
-  cherestapis.image.name: amisevsk/che-rest-apis:latest
+  cherestapis.image.name: amisevsk/che-rest-apis:v0.0.2
   che.webhooks.enabled: "false"
   che.default_routing_class: "basic"
